requirements.txt
# Core ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
statsmodels>=0.14.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
jax>=0.4.0
jaxlib>=0.4.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.0
ray[rllib]>=2.5.0

# Traditional ML
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.2.0

# Financial Data
yfinance>=0.2.0
alpaca-trade-api>=3.0.0
ccxt>=4.0.0
pandas-ta>=0.3.14b
TA-Lib>=0.4.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Database
psycopg2-binary>=2.9.0
redis>=4.5.0
pymongo>=4.4.0

# Data Processing
pyarrow>=12.0.0
polars>=0.18.0
dask>=2023.6.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.24.0

# Backtesting
backtrader>=1.9.0
zipline-reloaded>=3.0.0
vectorbt>=0.25.0

# Genetic Algorithms
deap>=1.3.0
pygad>=3.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
websockets>=11.0.0
schedule>=1.2.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Jupyter and Analysis
jupyter>=1.0.0
ipykernel>=6.24.0
nbformat>=5.9.0

# News and Sentiment
newspaper3k>=0.2.8
textblob>=0.17.0
transformers>=4.30.0

# Monitoring
prometheus-client>=0.17.0
grafana-client>=3.5.0

# Cloud and Storage
boto3>=1.26.0
google-cloud-storage>=2.9.0
azure-storage-blob>=12.16.0 